# Use faster disjointness check for markers